backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-environ==0.11.2
reportlab==4.0.7
openpyxl==3.1.2
django-filter==23.5
channels==4.0.0
channels-redis==4.1.0
django-storages==1.14.2
boto3==1.29.7

